Display chat usernames and error messages in bold
Chat applications often display metadata like usernames, timestamps, or
error messages in a way that is easy to visually distinguish.  This not
only makes it quicker for users to read the chat output but also
discourages malicious users from spoofing chat output by using the same
text as the client.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>