Remove redundant MainWindow::audioEnabled
The audioEnabled flag used by clients to decide whether or not to call
NJClient::AudioProc() does not seem necessary.  We only run an audio
streamer when audio should be enabled anyway, so there is no need for
this flag.  Furthermore, if NJClient is not connected to a server we
should only hear local channel monitoring - no metronome and obviously
no remote users.

This patch drops MainWindow::audioEnabled since we don't really need it.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>