Add PortAudio support
Add a PortAudio audioStreamer to replace the Linux-only ALSA
streamer in the Qt client.  This patch changes the build to link against
libportaudio and introduces a zero-configuration PortAudio streamer.

Note that we use local 8-bit character encoding for all PortAudio
strings.  In the future PortAudio will switch to UTF-8 but it currently
uses ASCII and the local encoding, see
http://www.assembla.com/spaces/portaudio/tickets/200.

Next we need a audio device configuration so that non-trivial use cases
are possible.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>