Use audio configuration settings
This patch extends the PortAudio streamer to take a host API (e.g.
WinMME, ALSA, CoreAudio) and input/output devices.  The streamer will
try to use those settings but fall back to defaults if they cannot be
activated.

Also wire up the QSettings for PortAudio configuration and the
PortAudioConfigDialog.

It is now possible to select sound cards!

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>