Show connect/disconnect status and error messages
This patch shows status notifications and error messages in the chat
textbox.  This makes it possible to tell when authentication has failed
or we have disconnected from the server.

The patch adds the ClientRunThread::statusChanged() signal which fires
when NJClient::GetStatus() changes.  Typically this happens when a state
transition happens from "disconnected" to "connected" or "connecting" to
"connecting failed".  Also ensure that we call Disconnect() when the
client enters an error state - this will re-enable the "Connect..." menu
item so the user can try again.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>