Show BPM/BPI not available when disconnected
When the client starts up or after disconnecting the BPM/BPI settings
take on default values.  It's nicer if we only update and display them
while connected to a server.

Take this opportunity to split beatInfoChanged() into two separate
signals for BPM and BPI.  This will make it easier to send BPI change
signals directly to the metronome bar in the future.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>