Add About... dialog
The About... dialog shows the program version number and includes links
to the Wahjam website, NINJAM, the GPLv2 license, as well as the git
commit that the binary was built from.

On systems without git or a missing PATH to git the commit ID will be
empty but the build will succeed.

Note that I have chosen version 0.9.0 as the current version.  The
intent is to release 1.0.0 soon.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>