Add MetronomeBar widget
This patch introduces a custom widget for showing the beats and current
interval.  The widget looks and operates similar to a progress bar but
is designed to draw each beat as a separate rectangle - something that
is hard to do with the standard QProgressBar.

Ikkei Shimomura (tea) <ikkei.shimomura@gmail.com> wrote a Python demo to
prototype this custom widget.  I implemented it in C++ based on his
example.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>