Show audio configuration dialog on first start
When launching for the first time it's important to select the right
sound card.  Therefore show the audio configuration dialog instead of
the connection dialog on first start.

Note that the "lastLaunchedVersion" setting may also come in handy if we
want to display a "What's new..." dialog on first launch after an update.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>