Add URL link support to chatAddLine()
Vote [+1] links are special-cased to append a URL link to the end of a
chat line.  Links could be used for other purposes in the future so it's
nice to support them directly in chatAddLine() and chatAddMessage().

This refactoring also eliminates the Vote [+1] dependency on
chatAddLine() cursor positioning.  It relied on the fact that
chatAddLine() does not include a newline at the end.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>