Keep a single global QSettings instance
QSettings uses a platform-specific mechanism to store application
settings, including the registry on Windows or .ini files on Linux.  It
is also possible to override the mechanism to avoid using the registry
on Windows, for example.  This is useful when building a portable app
that can run from a USB stick without installation.

In order to make it easy to override the QSettings mechanism in the
future, keep a single global QSettings instance.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>