Use QStateMachine for connection state
Several widget properties need to be updated when the client transitions
between connected and disconnected state.  Qt offers a state machine API
to implement this type of behavior in a more declarative way at a single
point in the source code.  This patch refactors the code to use a state
machine.

The new Connected() signal is added and complements the existing
Disconnected() signal.

Signed-off-by: Ikkei Shimomura (tea) <Ikkei.Shimomura@gmail.com>
Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>