Remove platform-specific clients
The Qt client is mature and stable enough for regular use.  Maintaining
the other, platform-specific clients is not feasible since any client
improvement needs to be reimplemented for each client.  It is also
difficult to improve core client code without making code changes to
platform-specific clients and testing them on all platforms.

For these reasons it is now time to remove the platform-specific clients
and rely solely on Qt and PortAudio cross-platform libraries.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>