Convert listener to QTcpServer
Use the Qt QTcpServer class to listen for connections instead of
JNL_Listen.  The advantage of QTcpServer is that it is event-driven and
does not need polling.

Since Net_Connection has not been converted to QTcpSocket yet we dup(2)
the socket file descriptor and then close the QTcpSocket.  The server
still uses JNL_Connection for client connections.  Future patches will
make it possible to use QTcpSocket instead of JNL_Connection.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>