Convert wsprintf() back to sprintf()
The jnetlib library #defined wsprintf to sprintf.  Once we switch to
QTcpSocket and drop jnetlib we will no longer have this macro.  Switch
back to sprintf() for now.

The long term plan for strings is to use QString so that encoding is
handled and the const char* vs char* warnings can be fixed.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>