Replace JNL_Connection with QTcpSocket
Convert the client connections from JNL_Connection to QTcpSocket.  This
is just the first step and still relies on polling.

The next step is to use signals/slots for event-driven I/O and
disconnect handling.

Note: after disconnect the remote address is no longer available and the
server will print an empty IP address in its log.  This will be fixed
once signals/slots are used for disconnect handling.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>