Convert wahjamsrv SessionArchive timer to QTimer
The wahjamsrv SessionArchive feature stores audio in session archive
directories so it can be mixed after the jam has finished.  Each session
archive directory contains the audio for a preconfigured number of
minutes.

This patch converts the session archive timer to QTimer.  Perhaps the
idle state can be eliminated in the future so the timer is only active
when authenticated users are logged in.  This could be done by turning
User_Group into a QObject and emitting signals for authentication
success and and log out.  The current implementation only triggers every
30 seconds though so there will not be much difference.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>