Adapt signal handlers to Qt event loop
This commit finally converts the main loop to app.exec().  The last
missing piece was the POSIX signal handlers that need to be adapted to
the Qt event loop.

Qt integration for POSIX signals is poor so we use ugly adapter code
that converts signals into writes on a socketpair(2) file descriptor.

The good news is that the server is now event-driven and only consumes
CPU when there is event processing work to be done!

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>