Use QTimer instead of ClientRunThread to call NJClient::Run()
ClientRunThread invokes NJClient::Run() every 20 milliseconds and
notifies MainWindow when state changes.  It turns out we can use QTimer
for the timer interval and get rid of the thread!  This makes life
easier because we only access NJClient from the Qt event loop and no
longer need a mutex.

The next step is to push the timer and state change signals down into
NJClient itself so MainWindow does not have to do so much babysitting.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>