Fix vote [+1] links in chat
The automatic vote [+1] links are broken because
QTextBrowser::openExternalLinks does not emit anchorClicked() when
enabled.  Disable both openExternalLinks and openLinks so that
QTextBrowser leaves all link handling up to the anchorClicked() handler.
When a [+1] link is clicked we send a vote chat message and when a web
link is clicked we open the URL.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>