Remove "Boost (+3dB)" checkbox
The "Boost (+3dB)" feature increases the playback volume of the
metronome or local channel.  The idea was to help the metronome or
user's instrument cut through the mix.

This patch removes the feature because it is rarely used and can confuse
users who believe it increases their remote volume.  Only the local
playback volume is increased, remote users do not hear louder volume.

Instead of manipulating local volume, jams work best when each user
tries to fit into the mix.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>