Remove audio buffer files
Let NJClient delete temporary .ogg files.  It is useful to keep them
around for debugging or post-jam mixing.  For now it's most
user-friendly to delete them so we don't silently fill up the user's
harddisk.  In the future we may introduce a setting where the user can
keep jams saved.

Also fix the logic of printing a "Disconnected" message.  It should not
depend on whether or not to delete the session directory.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>