Fix initializer list order in njclient.cpp
It is good practice to keep initializer lists ordered the same way as
the member variables are declared.  This is because C++ constructors
initialize in order of member variable declarations, regardless of the
initializer list order.  Keeping the order consistent prevents confusing
constructor bugs.

This patch fixes a gcc 4.7.1 warning due to initializer list order.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>