Rename "qtclient" binary to "wahjam"
Support for the legacy clients (curses, Windows, Mac OS X) has been gone
for a while.  The only client is the cross-platform Qt client.  Call the
binary "wahjam" so packaging scripts don't need to duplicate this step
after the binary has been built.

Note we still use the qtclient.pro project file.  Eventually we should
move to a global project file for wahjam, wahjamsrv, and other build
targets.  In the meantime there is no need to rename the project file.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>