Add JammrServerBrowser for Jammr REST API server list
The Jammr API returns a server list from http://<api-root>/livejams/.
We use XML format since Qt 4 does not ship with a JSON module.

Note that for Jammr we don't want to display the server connection
details since they are unlikely to interest the user.  Instead we
display the server's topic along with the usual tempo and user list
information.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>