Add JammrConnectDialog
The Jammr platform has single sign-on and therefore doesn't require the
user to give a username/password for each server they connect to.  There
are also no anonymous users.

These differences to the NINJAM connection model make it worth having a
separate "Connect to server..." dialog for Jammr.

Note that ConnectDialog::loadServerList() takes a URL but
JammrConnectDialog takes the URL through its constructor.  This
difference is because JammrConnectDialog also needs the API URL in the
future for creating private jams.  This functionality is added in a
later patch.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>