Remove broken Net_Message refcount
Reference counting is easy to get wrong and the code does not use it
consistently.  Luckily reference counting is not necessary to manage
Net_Message memory allocation.  It can be replaced with a small change
to the Net_Connection::Send() prototype:

  int Send(Net_Message *msg, bool deleteAfterSend = true)

Callers wishing to broadcast or send a message without deleting it can
set deleteAfterSend to false.  This allows them to delete the message
manually later.  This is used in the server's usercon.cpp code.

With this change in place, it's possible to drop Net_Message reference
counting completely.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>