Emit Net_Connection::disconnected() even on connection refused
Socket errors that happen before the connection is established are
silent because the Net_Connection::disconnected() signal is not being
emitted.

This patch fixes the missing connection refused behavior.  For example,
if a connection to the wrong IP address is attempted the client will now
display a connection refused error.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>