Add Jammr protocol version
The Jammr protocol is based on NINJAM but is incompatible because all
text is UTF-8 encoded whereas the encoding is undefined for NINJAM
clients.  Additional protocol extensions are being developed and make it
necessary to use a different protocol version number from NINJAM to
prevent collisions in the future.

The number 0x80000000 was chosen to give NINJAM and other variants
plenty of room.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>