Remove ChannelTree broadcast check state
Each local channel has a broadcast state which controls whether or not
the audio is uploaded to the server.  There is now a global broadcast
button ("Xmit") to replace per-channel broadcast checkboxes.

Remove the ChannelTree broadcast checkboxes because they can be
confusing and are no longer needed.  Currently only one local channel is
supported so there is no disadvantage to exclusively using the global
broadcast button.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>