Add connecting state to MainWindow state machine
Keep track of when the client is connecting to the server.  During this
time it is possible to disconnect.

This fixes a bug where it is possible to connect to a new jam session
when the client is already connecting.  When this happens the audio
stream is opened twice and the metronome runs at double speed!

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>