Fix win32 libcommon.a dependencies
qmake does not take care of sub-project dependencies automatically.
qtclient.pro and server.pro manually specify dependencies on
libcommon.a.  The path to libcommon.a is different when cross-compiling
for win32 since there is a release/ or debug/ subdirectory.

Capture the ugly details of the libcommon.a dependency in a qmake
include file which qtclient.pro and server.pro can include.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>