Add JammrUpdateChecker for automatic update polling
The update checker fetches a URL and if the returned version string does
not match the current version, it pops up a dialog prompting the user to
download the update.

Some operating systems have better solutions, like package managers
under Linux.  For now we use the update checker on all platforms until
jammr installers and updates are customized for different operating
systems.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>