Augment sample rate autoselection with manual controls
Previously the PortAudio AudioStreamer attempted to autoselect good
sample rate and latency settings.  This worked fine on some systems but
ultimately left some users with non-functioning audio.

Move the autoselection code into PortAudioConfigDialog and add manual
controls to set the sample rate and latency.  Also check the audio
settings so that the "Apply" button is only enabled when the system
actually supports these settings.

Note that we still do autoselection by default when the user selects a
device, but they now have a change to manually select the sample rate
and latency.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>