Convert PortAudioConfigDialog to PortAudioSettingsPage
Move PortAudio configuration into the Settings dialog.  A few changes
are necessary because the old dialog lifecycle does not exist when
converted to a Settings dialog page:

The Apply/Cancel buttons are removed.  Therefore all changes have
immediate effect.  This is similar to the Google Chrome settings dialog.

The old dialog disabled the Apply button if the sound devices did not
support the current settings.  The new Settings dialog page has a QLabel
that tells the user whether or not these settings will work.  Invalid
settings can be saved but the an error dialog will be displayed before
bringing the user back to the audio settings.

Saving/loading QSettings has been adjusted so that settings are loaded
on startup and saved each time the Settings dialog is closed.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>