Introduce EffectPlugin base class
Extract a base class from VSTPlugin so new types of effect plugins
besides VST can be implemented.

The VSTProcessor no longer needs to know about VST or any specific
plugin standard.  Instead it uses the EffectPlugin interface and is
renamed EffectProcessor.

Note that the Settings dialog UI is still hardcoded for VST plugins.
This will be fixed in another commit.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>