Add "Show log" help menu item
The log.txt file is located in the QDesktopServices::DataLocation
directory.  This directory is dependent on the operating system release
and can be hard for users to find.

Add a help menu item that opens the log file.  This makes it easy for
users to email or pastebin the log when they encounter problems.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>