Switch to QUrlQuery class from QUrl::addQueryItem()
The QUrl::addQueryItem() method was removed in favor of the QUrlQuery
class in Qt5.  QUrlQuery represents a URL query string and allows access
to the key-value pairs.

In simple cases we can directly call QUrl::setQuery("key=value").  In
more complex cases we need to build up a QUrlQuery object and pass it to
QUrl::setQuery().

Signed-off-by: Ikkei Shimomura (tea) <Ikkei.Shimomura@gmail.com>
Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>