Fix incorrect scratch buffer sizes in EffectProcessor
The EffectProcessor keeps two sets of working buffers around.  They are
used for inputs and outputs, respectively.  When there is more than one
EffectPlugin in a chain, the processing function swaps these two buffer
sets as it lets each plugin generate audio.

This logic was broken because we distinguished between maximum number of
inputs and outputs for a plugin.  Since the buffer sets are swapped
after each iteration of the processing loop, this can lead to
out-of-bound memory accesses if the number of inputs and outputs differ.

Simplify the logic by just keeping a maximum number of inputs or outputs
and allocating two sets of these buffers.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>