Work around NULL QTextCodec::codecForLocale()
QTextStream uses a codec to encode/decode Unicode characters.  On Qt5
Windows builds QTextCodec::codecForLocale() seems to return NULL and
results in a crash inside Qt.

Set the default codec to UTF-8 as a workaround.  Eventually this will go
away when the underlying problem gets fixed in Qt or mxe packages.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>