Honor PortAudio's UTF-8 encoding
Host API and device names are UTF-8 encoded by PortAudio.  The
exceptions are the WinMME and DirectSound host APIs, but that is a bug
in PortAudio that is being fixed.

Users have reported that WDM KS devices with non-ASCII names are not
being recognized.  This patches fixes this.

Although WinMME and DirectSound may behave differently after this patch,
they were always broken.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>