Fix "Send" button on startup
When the "Send" button was disabled last time jammr was shut down, it
appears disabled on startup too but audio is actually being transmitted.

This happens because QToolButton->setChecked() does not emit a toggled()
signal if the new state matches the old state.
MainWindow::setupStatusBar() was making this assumption and therefore
didn't disable transmit properly.

Interestingly the metronome suffers the same problem but already had a
fix.  Consolidate both into the setupStatusBar() function.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>