Hide metronome bar when metronome is disabled
Some users may think that they need to align to the beginning of the
metronome bar animation even when the metronome has been disabled.  Make
it clear that the metronome animation is not relevant when the metronome
has been disabled.

It has also been reported that the metronome bar animation is
distracting when not in use.

The simple solution is to hide the bar when the metronome is disabled.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>