Fix focus in jammr Connect dialog
The jammr Connect dialog had input focus on the "New jam" button.  This
made it difficult for keyboard or screen reader users to press the
"Connect" button (which is disabled until the list of jams is loaded
into the server browser).

Put the focus on the server browser, with the "Connect" button as the
next widget and finally the "New jam" button.  Most users connect to
existing jams and do not create new jams.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>