cliplogcvt: drop libresample from pkgconfig
The libresample pkgconfig files are shipped by some distros (Debian and
Fedora) but aren't part of the upstream source tarball.  Unfortunately
this means pkgconfig support isn't available everywhere and the build
would fail.

Luckily the compile-time requirements of libresample are so simple that
we can hard-code them instead of using pkgconfig.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>