Send MIDI Beat Clock messages with explicit timestamps
Specify the MIDI output latency and send MIDI messages with timestamps.
This improves the MIDI Beat Clock accuracy because we only update it for
each multi-millisecond audio buffer. Before the MIDI Beat Clock messages
could be off by several milliseconds and this made it hard for external
devices and applications to sync to the tempo.

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>