Add unlock button to audio and MIDI settings pages
Audio and MIDI settings cannot be changed while connected to a jam
session. Users are often confused because the widgets are disabled to
prevent changes.

Show an unlock button that disconnects from the jam session. This makes
it clear that disconnecting is required in order to make changes but
saves the user a trip to File | Disconnect.

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>