Replace MIDI queues with direct PortMidi access
The idea of ConcurrentQueue<PmEvent> was to handle MIDI processing in a
dedicated thread. This would allow the audio processing thread to be
real-time.

After auditing PortMidi's Windows, Mac, and Linux code I've come to the
conclusion that its timing properties are sufficient to be used directly
from the audio processing thread.

Directly calling PortMidi eliminates the 1 millisecond delay that the
MIDI processing thread imposed. This should help with the timing of MIDI
events.

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>