Specify fixed PortAudio buffer size
When no fixed buffer size is specified PortAudio invokes the audio
processing callback with partial buffers when it cross the end of the
ring buffer. Some VST plugins like S-Gear do not respond well to buffer
size changes.

Use a different approach: ask PortAudio to used a fixed buffer size but
if the actual latency reported for the device is different once the
stream has started, try starting it again with the fixed buffer size.

This basically makes a small adjustment to the user-specified latency
value in order to find the optimal buffer size that PortAudio can
provide.

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>