Use PortAudio stream as time reference for PortMidi
The key to synchronizing MIDI to audio streams is to use the output DAC
time for scheduling MIDI events and the current audio stream time as the
time reference.

This way PortMidi is able to send MIDI messages with correct timestamps
independent of when the audio processing callback was invoked.

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>