Add "Input noise protection" to prevent feedback
First-time users may have forgotten to use headphones when recording
through a microphone or the input device may be set to the built-in
microphone by mistake. These things can cause unwanted feedback that
disrupts the jam.

Add a simple loud noise detection algorithm and add an audio settings
checkbox for "Input noise protection". This feature is enabled by
default but can be disabled if users find the algorithm is producing
false positives.

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>