Add wahjamsrv VotingThreshold 1 mode
Add a special case for VotingThreshold 1. Normally this would require at
least 1% of users to vote for a BPI/BPM change. 1% is so small that this
threshold is effectively useless.

Change the meaning of VotingThreshold 1 to require just 1 additional
vote from another user. This is a useful mode because jams with more
than 6 users often have trouble reaching the voting threshold because
users are too busy jamming to vote. Most of the time voting is just
needed to prevent accidental or malicious BPI/BPM changes.

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>