Add stretch to EffectsSettingsPage button layout
The buttons should appear top-to-bottom with empty space if the window
is big enough. Currently they are equally spaced, which looks odd and
depends on the size of other widgets in the settings dialog.

Add stretch so the buttons stay in a vertical column at the top of the
dialog with fixed spacing.

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>